Move ts files relating to style in the same folder
