+++ src/style/FileItemStyle.tsx
