+++ src/style/GitPanelStyle.tsx
