+++ src/style/GitStageStyle.tsx
