+++ src/style/GitWidgetStyle.tsx
