+++ src/style/HistorySideBarStyle.tsx
