+++ src/style/NewBranchBoxStyle.tsx
