+++ src/style/SinglePastCommitInfoStyle.tsx
